Completed coap-device and  test

+++ CoAP-network/coap-device/CO2-dispenser-var.h
@@ -1,5 +1,5 @@
 extern bool co2_to_be_filled;
-extern float co2_tank_level;
-extern float co2_value;
+extern int co2_tank_level;
+extern int co2_value;
 extern bool co2_flow;
 extern bool co2_to_stop;
