+++ CoAP-network/coap-device/osmotic-water-device-var.h
@@ -1,3 +1,3 @@
 extern bool water_tank_to_be_filled;
-extern float water_tank_level;
+extern int water_tank_level;
 extern bool water_tank_to_stop;
